requirements.txt
aerich==0.5.8
aiosqlite==0.17.0
asgiref==3.4.1
asyncpg==0.24.0
click==8.0.1
databases==0.5.2
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.68.1
fastapi-pagination==0.9.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
iso8601==0.1.16
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dotenv==0.19.0
pytz==2021.3
PyYAML==5.4.1
SQLAlchemy==1.4.25
starlette==0.14.2
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
